+++ sample/test/klite/jdbc/JdbcExtensionsTest.kt
@@ -111,6 +111,19 @@ open class JdbcExtensionsTest: TempTableDBTest() {
     expect(db.call("length", "hello", returnSqlType = Types.INTEGER)).toEqual(5)
   }
 
+  @Test fun `postgres advisory locks`() {
+    val ok = "Hello"
+    expect(db.unlock(ok)).toEqual(false)
+    try {
+      expect(db.tryLock(ok)).toEqual(true)
+      expect(db.tryLock(ok)).toEqual(true)
+    } finally {
+      expect(db.unlock(ok)).toEqual(true)
+      expect(db.unlock(ok)).toEqual(true)
+      expect(db.unlock(ok)).toEqual(false)
+    }
+  }
+
   @Test fun `postgres notify and listen`() = runTest {
     val channel = Channel<String>(UNLIMITED)
     val reader = thread {