+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# Unreleased
+* jdbc: let tryLock/unlock not eat exceptions; also log if they return false
+
 # 2.0.3
 * server: AppScope.async will now log any exceptions to avoid silent failures
 * json: use explicit buffering in JsonParser and JsonReader to speed up by ~25% (avoids virtual calls and char-int conversions)